+++ Framework/Built_In_Automation/Sequential_Actions/action_declarations/common.py
@@ -111,8 +111,12 @@ declarations = (
     {"name": "xml to json",                                  "function": "xml_to_json",                           "screenshot": "none" },
 
     # AI actions
-    {"name": "classifier ai",                                "function": "classifier_AI",                          "screenshot": "none" },
+    {"name": "classifier ai",                                "function": "classifier_AI",                         "screenshot": "none" },
 
+    {"name": "connect to S3",                                "function": "connect_to_S3",                         "screenshot": "none" },
+    {"name": "upload to S3",                                 "function": "upload_to_S3",                          "screenshot": "none" },
+    
+    {"name": "text to speech",                               "function": "text_to_speech",                        "screenshot": "none" },
 ) # yapf: disable
 
 module_name = "common"